requirements.txt
Django==1.10.5
Pillow==3.4.2
django-simple-captcha==0.5.3
djangorestframework==3.4.7
mysqlclient==1.3.9
pytz==2016.7